refactor(sdk): extract PhaseExecutor interface for multi-provider support
Define PhaseExecutor interface with name, tier, and runPhase() method.
Create ClaudeAgentExecutor (Tier 1) implementing it, wrapping current
query() logic. All execution functions now accept a PhaseExecutor as
first parameter instead of importing runPhase directly.

This is the abstraction seam for future providers. Zero behavioral
change — all 212 tests pass, build succeeds, CLI works.

Chain: CLI creates executor → executeSingleItem(executor, ...) →
executor.runPhase() → SDK query()

Refs: docs/analysis/20260402_spike_multi-provider-hybrid-architecture.md
