+++ src/execution/single-item.test.ts
@@ -4,15 +4,15 @@ import {
   type SingleItemOptions,
   type SingleItemResult,
 } from "./single-item.js";
-import type { PhaseResult } from "../core/phase-executor.js";
+import type { PhaseExecutor, PhaseResult, PhaseOptions } from "../core/phase-executor.js";
 
-// Mock runPhase
-vi.mock("../core/phase-executor.js", () => ({
-  runPhase: vi.fn(),
-}));
+const mockRunPhase = vi.fn<(phase: string, input: string, options?: PhaseOptions) => Promise<PhaseResult>>();
 
-
-import { runPhase } from "../core/phase-executor.js";
+const mockExecutor: PhaseExecutor = {
+  name: "mock",
+  tier: 1,
+  runPhase: mockRunPhase,
+};
 
 function mockPhaseResult(overrides?: Partial<PhaseResult>): PhaseResult {
   return {
@@ -34,42 +34,42 @@ describe("executeSingleItem", () => {
 
   it("runs phases from discover through done", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    const result = await executeSingleItem("user auth", {
+    const result = await executeSingleItem(mockExecutor,"user auth", {
       fromPhase: "discover",
       throughPhase: "done",
     });
 
     // Assert — should call runPhase for each phase
-    expect(vi.mocked(runPhase).mock.calls.length).toBe(7);
+    expect(mockRunPhase.mock.calls.length).toBe(7);
     expect(result.exitCode).toBe(0);
   });
 
   it("respects fromPhase and throughPhase range", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("docs/backlog/P0-item.md", {
+    await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "design",
       throughPhase: "deliver",
     });
 
     // Assert — only design and deliver
-    const phases = vi.mocked(runPhase).mock.calls.map((c) => c[0]);
+    const phases = mockRunPhase.mock.calls.map((c) => c[0]);
     expect(phases).toEqual(["design", "deliver"]);
   });
 
   it("accumulates total cost across phases", async () => {
     // Arrange
-    vi.mocked(runPhase)
+    mockRunPhase
       .mockResolvedValueOnce(mockPhaseResult({ costUsd: 0.05 }))
       .mockResolvedValueOnce(mockPhaseResult({ costUsd: 0.10 }));
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "design",
       throughPhase: "deliver",
     });
@@ -80,32 +80,32 @@ describe("executeSingleItem", () => {
 
   it("passes session IDs for cross-phase resume", async () => {
     // Arrange
-    vi.mocked(runPhase)
+    mockRunPhase
       .mockResolvedValueOnce(mockPhaseResult({ sessionId: "sess-discover" }))
       .mockResolvedValueOnce(mockPhaseResult({ sessionId: "sess-define" }));
 
     // Act
-    await executeSingleItem("user auth", {
+    await executeSingleItem(mockExecutor,"user auth", {
       fromPhase: "discover",
       throughPhase: "define",
     });
 
     // Assert — second call should get resume from first
-    const secondCall = vi.mocked(runPhase).mock.calls[1];
+    const secondCall = mockRunPhase.mock.calls[1];
     const opts = secondCall[2];
     expect(opts?.resumeSessionId).toBe("sess-discover");
   });
 
   it("detects APPROVED verdict from discern phase", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(
+    mockRunPhase.mockResolvedValue(
       mockPhaseResult({
         output: "Review complete. Verdict: APPROVED.",
       }),
     );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
     });
@@ -117,14 +117,14 @@ describe("executeSingleItem", () => {
 
   it("detects BLOCKED verdict and exits 1", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(
+    mockRunPhase.mockResolvedValue(
       mockPhaseResult({
         output: "Critical issues found. Verdict: BLOCKED.",
       }),
     );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
     });
@@ -136,7 +136,7 @@ describe("executeSingleItem", () => {
 
   it("handles CHANGES_REQUESTED with review cycle", async () => {
     // Arrange — first discern requests changes, deliver fixes, second discern approves
-    vi.mocked(runPhase)
+    mockRunPhase
       .mockResolvedValueOnce(
         mockPhaseResult({ output: "Verdict: CHANGES REQUESTED. Fix the tests." }),
       )
@@ -148,38 +148,38 @@ describe("executeSingleItem", () => {
       );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
       reviewCycles: 2,
     });
 
     // Assert — should have done: discern → deliver → discern
-    const phases = vi.mocked(runPhase).mock.calls.map((c) => c[0]);
+    const phases = mockRunPhase.mock.calls.map((c) => c[0]);
     expect(phases).toEqual(["discern", "deliver", "discern"]);
     expect(result.verdict).toBe("APPROVED");
   });
 
   it("stops review cycles at max limit", async () => {
     // Arrange — always requests changes
-    vi.mocked(runPhase).mockResolvedValue(
+    mockRunPhase.mockResolvedValue(
       mockPhaseResult({ output: "Verdict: CHANGES REQUESTED." }),
     );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
       reviewCycles: 1,
     });
 
     // Assert — discern once, deliver once (review cycle), discern again = 3 calls max
-    expect(vi.mocked(runPhase).mock.calls.length).toBeLessThanOrEqual(3);
+    expect(mockRunPhase.mock.calls.length).toBeLessThanOrEqual(3);
   });
 
   it("exits 1 when review cycle results in BLOCKED", async () => {
     // Arrange — discern requests changes, deliver fixes, second discern blocks
-    vi.mocked(runPhase)
+    mockRunPhase
       .mockResolvedValueOnce(
         mockPhaseResult({ output: "Verdict: CHANGES REQUESTED." }),
       )
@@ -191,7 +191,7 @@ describe("executeSingleItem", () => {
       );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
       reviewCycles: 2,
@@ -204,12 +204,12 @@ describe("executeSingleItem", () => {
 
   it("exits 1 when phase exhausts turns", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(
+    mockRunPhase.mockResolvedValue(
       mockPhaseResult({ exhausted: true }),
     );
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
@@ -220,10 +220,10 @@ describe("executeSingleItem", () => {
 
   it("collects phase results for logging", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    const result = await executeSingleItem("docs/backlog/P0-item.md", {
+    const result = await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "design",
       throughPhase: "deliver",
     });
@@ -242,40 +242,40 @@ describe("executeSingleItem cost logging", () => {
 
   it("passes hooks to runPhase when logDir is set", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("docs/backlog/P0-item.md", {
+    await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
       logDir: "/tmp/test-logs",
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hooks).toBeDefined();
     expect(phaseOpts?.hooks?.onResult).toBeTypeOf("function");
   });
 
   it("passes onToolUse but not onResult when logDir is absent", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("docs/backlog/P0-item.md", {
+    await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
 
     // Assert — onToolUse always present (artifact tracking), onResult absent
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hooks?.onToolUse).toBeTypeOf("function");
     expect(phaseOpts?.hooks?.onResult).toBeUndefined();
   });
 
   it("passes hooks to review-cycle runPhase calls too", async () => {
     // Arrange — discern returns CHANGES_REQUESTED, then APPROVED
-    vi.mocked(runPhase)
+    mockRunPhase
       .mockResolvedValueOnce(
         mockPhaseResult({ output: "Verdict: CHANGES REQUESTED." }),
       )
@@ -285,7 +285,7 @@ describe("executeSingleItem cost logging", () => {
       );
 
     // Act
-    await executeSingleItem("docs/backlog/P0-item.md", {
+    await executeSingleItem(mockExecutor,"docs/backlog/P0-item.md", {
       fromPhase: "discern",
       throughPhase: "discern",
       reviewCycles: 2,
@@ -293,7 +293,7 @@ describe("executeSingleItem cost logging", () => {
     });
 
     // Assert — all 3 calls (discern, deliver, discern) should have hooks
-    for (const call of vi.mocked(runPhase).mock.calls) {
+    for (const call of mockRunPhase.mock.calls) {
       const phaseOpts = call[2];
       expect(phaseOpts?.hooks?.onResult).toBeTypeOf("function");
     }
@@ -307,32 +307,32 @@ describe("executeSingleItem auth mode", () => {
 
   it("passes authMode to runPhase when set", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("item.md", {
+    await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
       authMode: "oauth",
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.authMode).toBe("oauth");
   });
 
   it("does not pass authMode when not set", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("item.md", {
+    await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.authMode).toBeUndefined();
   });
 });
@@ -344,22 +344,22 @@ describe("executeSingleItem artifact tracking", () => {
 
   it("passes onToolUse hook to runPhase", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("item.md", {
+    await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hooks?.onToolUse).toBeTypeOf("function");
   });
 
   it("records Write tool_use file paths as artifacts", async () => {
     // Arrange — simulate SDK PostToolUse hook firing
-    vi.mocked(runPhase).mockImplementation(async (_phase, _input, opts) => {
+    mockRunPhase.mockImplementation(async (_phase, _input, opts) => {
       opts?.hooks?.onToolUse?.({
         toolName: "Write",
         toolInput: { file_path: "src/index.ts", content: "..." },
@@ -368,7 +368,7 @@ describe("executeSingleItem artifact tracking", () => {
     });
 
     // Act
-    const result = await executeSingleItem("item.md", {
+    const result = await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
@@ -379,7 +379,7 @@ describe("executeSingleItem artifact tracking", () => {
 
   it("records Edit tool_use file paths as artifacts", async () => {
     // Arrange
-    vi.mocked(runPhase).mockImplementation(async (_phase, _input, opts) => {
+    mockRunPhase.mockImplementation(async (_phase, _input, opts) => {
       opts?.hooks?.onToolUse?.({
         toolName: "Edit",
         toolInput: { file_path: "src/cli.ts", old_string: "a", new_string: "b" },
@@ -388,7 +388,7 @@ describe("executeSingleItem artifact tracking", () => {
     });
 
     // Act
-    const result = await executeSingleItem("item.md", {
+    const result = await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
@@ -399,7 +399,7 @@ describe("executeSingleItem artifact tracking", () => {
 
   it("deduplicates artifacts across phases", async () => {
     // Arrange
-    vi.mocked(runPhase).mockImplementation(async (_phase, _input, opts) => {
+    mockRunPhase.mockImplementation(async (_phase, _input, opts) => {
       opts?.hooks?.onToolUse?.({
         toolName: "Write",
         toolInput: { file_path: "src/index.ts", content: "..." },
@@ -408,7 +408,7 @@ describe("executeSingleItem artifact tracking", () => {
     });
 
     // Act
-    const result = await executeSingleItem("item.md", {
+    const result = await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "design",
       throughPhase: "deliver",
     });
@@ -420,7 +420,7 @@ describe("executeSingleItem artifact tracking", () => {
 
   it("ignores non-Write/Edit tool uses", async () => {
     // Arrange
-    vi.mocked(runPhase).mockImplementation(async (_phase, _input, opts) => {
+    mockRunPhase.mockImplementation(async (_phase, _input, opts) => {
       opts?.hooks?.onToolUse?.({
         toolName: "Bash",
         toolInput: { command: "ls" },
@@ -433,7 +433,7 @@ describe("executeSingleItem artifact tracking", () => {
     });
 
     // Act
-    const result = await executeSingleItem("item.md", {
+    const result = await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
@@ -444,7 +444,7 @@ describe("executeSingleItem artifact tracking", () => {
 
   it("coexists with cost logger hooks", async () => {
     // Arrange
-    vi.mocked(runPhase).mockImplementation(async (_phase, _input, opts) => {
+    mockRunPhase.mockImplementation(async (_phase, _input, opts) => {
       opts?.hooks?.onToolUse?.({
         toolName: "Write",
         toolInput: { file_path: "src/new.ts", content: "..." },
@@ -453,7 +453,7 @@ describe("executeSingleItem artifact tracking", () => {
     });
 
     // Act
-    const result = await executeSingleItem("item.md", {
+    const result = await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
       logDir: "/tmp/logs",
@@ -461,7 +461,7 @@ describe("executeSingleItem artifact tracking", () => {
 
     // Assert — both hooks should fire: artifacts tracked AND cost logger wired
     expect(result.artifacts).toContain("src/new.ts");
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hooks?.onResult).toBeTypeOf("function");
     expect(phaseOpts?.hooks?.onToolUse).toBeTypeOf("function");
   });
@@ -474,32 +474,32 @@ describe("executeSingleItem hasContextDir", () => {
 
   it("passes hasContextDir to runPhase when set", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("item.md", {
+    await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
       hasContextDir: true,
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hasContextDir).toBe(true);
   });
 
   it("does not set hasContextDir when not provided", async () => {
     // Arrange
-    vi.mocked(runPhase).mockResolvedValue(mockPhaseResult());
+    mockRunPhase.mockResolvedValue(mockPhaseResult());
 
     // Act
-    await executeSingleItem("item.md", {
+    await executeSingleItem(mockExecutor,"item.md", {
       fromPhase: "deliver",
       throughPhase: "deliver",
     });
 
     // Assert
-    const phaseOpts = vi.mocked(runPhase).mock.calls[0][2];
+    const phaseOpts = mockRunPhase.mock.calls[0][2];
     expect(phaseOpts?.hasContextDir).toBeUndefined();
   });
 });
